Vai dar merge

+++ src/main/java/lapr/project/controller/PositionalMessageController.java
@@ -2,7 +2,7 @@ package lapr.project.controller;
 
 import lapr.project.model.Company;
 import lapr.project.model.Ship;
-import lapr.project.utils.auth.domain.store.ShipStore;
+import lapr.project.model.stores.ShipStore;
 
 public class PositionalMessageController {
 
@@ -15,9 +15,10 @@ public class PositionalMessageController {
     }
 
 
-    public boolean getPositionalMessages(long mmsi) {
+    public boolean getPositionalMessages(int mmsi) {
 
         Ship ship = shipStore.getShipByMMSI(mmsi);
+
 
         if (ship == null) return false;
